+++ src/event.h
@@ -46,21 +46,21 @@ struct local_private {
 
 	struct ro_remote *current_send_remote; /* We are currently sending to this remote */
 	struct ro_remote *current_receive_remote; /* We are currently receiving from this remote */
-	uint32_t remaining_bytes;	  /* We need to send this many bytes */
+	uint16_t remaining_bytes;	  /* We need to send this many bytes */
 	size_t partial_bytes;		  /* But before that, we only wrote this many bytes for the header */
 
-	uint32_t send_serial;	 /* Current serial number for sending */
-	uint32_t receive_serial; /* Current serial number for receiving */
+	uint16_t send_serial;	 /* Current serial number for sending */
+	uint16_t receive_serial; /* Current serial number for receiving */
 };
 
 struct remote_private {
 	struct event *read;
 	struct event *write;
 
-	uint32_t partial_header[2]; /* Partial received header */
+	uint16_t partial_header[2]; /* Partial received header */
 	size_t partial_bytes;	    /* Size of partially received header */
 
-	uint32_t remaining_bytes; /* We need to receive this many bytes */
+	uint16_t remaining_bytes; /* We need to receive this many bytes */
 };
 
 static inline void
